requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2>=3.1.0

# Database and Storage
supabase>=2.22.0,<3.0.0
psycopg2-binary==2.9.9

# OpenAI (for embeddings and question generation)
openai>=1.12.0,<2.0.0

# HTTP Requests
# Note: httpx version is managed by supabase dependency
requests==2.31.0

# Configuration and Environment
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
